+++ src/MetaheuristicsPlatform/Algorithms/GreyWolf/GreyWolfOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.GreyWolf;
+
+public static class GreyWolfOptimizerReferences
+{
+    public static ScientificReference MirjaliliMirjaliliLewis2014 { get; } =
+        new(
+            "Mirjalili, Mirjalili & Lewis",
+            2014,
+            "Grey Wolf Optimizer",
+            "Advances in Engineering Software 69, 46-61",
+            "10.1016/j.advengsoft.2013.12.007");
+}